Move setJourneyDetails middleware
