+++ src/modules/form/state/middleware.js
@@ -97,21 +97,7 @@ const setFormDetails = (req, res, next) => {
   next()
 }
 
-const setJourneyDetails = (journey, currentStep, currentStepId) => {
-  return (req, res, next) => {
-    res.locals.journey = {
-      currentStep,
-      currentStepId,
-      ...journey,
-      key: getFullRoute(req.baseUrl, journey.steps[0]),
-    }
-
-    next()
-  }
-}
-
 module.exports = {
-  setJourneyDetails,
   validateState,
   updateStateData,
   updateStateBrowseHistory,
